Rename component to GradientButton
